[image-editor-sizing] move utils into image folder
